Cursor bug resolution
Realized that the cursor bug fix had broken the ability to rotate
objects. Reverted the cursor bug fix. Re-fixed the bug differently, so
that now instead of the bug where the object doesn't move with the
cursor, we have a different bug wherein when the view has been rotated,
selecting an object will cause it to "jump" in rotation. We had to
choose between those two bugs, as the fix for one breaks the other. We
picked this version, where objects "jump" in rotation upon selection
when the view has been rotated.